html/browsers/browsing-the-web/navigating-across-documents/007.html

<!doctype html>
<link rel="help" href="https://html.spec.whatwg.org/multipage/browsing-the-web.html#navigating-across-documents">
<title>Link with onclick that sets location to "javascript: url" is aborted by href navigation</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe id="test" name="test"></iframe>
<a target="test" onclick="document.querySelector('#test').contentWindow.location = 'javascript:`<script>self.postMessage(&quot;click&quot;, &quot;*&quot;);</script>`';" href="href.html">Test</a>
<script>
var test = async_test();

test.step(function() {
  document.querySelector("a").click();
});

window.onmessage = test.step_func_done(function(e) {
  assert_equals(e.data, "href");
});
</script>

<!--
1. This test creates an iframe, and an anchor with both an `onclick` and `href`.
2. The test "clicks" the anchor, triggering the `onclick` handler operation,
    which will attempt to set the `location` of the iframe to a "javascript: url",
    that creates a script element containing a `postMessage` call, which will
    send a message "click".
3. Navigating the iframe to the anchor's `href` will abort the previous operation.
4. If the semantics are implemented correctly, a message "href"
    from "href.html" will be received.

-->
